Move DefaultImageDataRepository to production target
